// frontend/vite-project/src/types/DataTypes.ts
export interface AddItemInterface {
    categoryId: number;
    productName: string;
    productContent: string;
    paymentLink: string;
    price: number;
    deliveryCharge: number;
    quantity: number;
}
// ...
export type UploadImage = {
    file: File;
    type: string;
};